audio input boost support freephonie network audio device removeal should be ok
git-svn-id: http://svn.assembla.com/svn/symphonie/trunk@54 57038eb5-193d-0410-9cfd-c93977fb03f4

+++ pjproject-0.5.10.4/pjlib/src/pj/ioqueue_common_abs.c
@@ -848,7 +848,7 @@ PJ_DEF(pj_status_t) pj_ioqueue_sendto( pj_ioqueue_key_t *key,
     /*
      * Check that address storage can hold the address parameter.
      */
-    PJ_ASSERT_RETURN(addrlen <= sizeof(pj_sockaddr_in), PJ_EBUG);
+    PJ_ASSERT_RETURN(addrlen <= (int)sizeof(pj_sockaddr_in), PJ_EBUG);
 
     /*
      * Schedule asynchronous send.
